chore(backend): drop tests/support/, use real stub type for client responses
